Fail to do reasonable-sized commits again (aka I want modern C++)
The engine had been written way before post-C++11 era, in C++98/03.
And we've still been doing lots of C-style mess, although modern C++
is more like Rust.
I just wanted to start making new code more C++ and stop doing mess.
We can't rewrite these 1.5 mln locs right away, but we can write new
code properly.

So pull GSL, promote a bunch of compiler warnings to errors, fix them.
Introduce the new default namespace, 'xr', for engine code. Start
wrapping C includes into separate namespaces.
Start rewriting own crappy code. The first piece is _random.cpp -- now
it uses unique_ptr instead of raw pointers, own namespace etc.

Switch -march from Haswell to Skylake, it's been 10 years already.
Skylake optimization ensures we have hardware prefetching and more.

+++ ogsr_engine/xrGame/object_handler_planner_missile.cpp
@@ -21,10 +21,10 @@ void CObjectHandlerPlanner::add_evaluators(CMissile* missile)
 {
     u16 id = missile->ID();
     // dynamic state properties
-    add_evaluator(uid(id, eWorldPropertyHidden), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, CMissile::eHidden));
-    add_evaluator(uid(id, eWorldPropertyThrowStarted), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, CMissile::eThrowStart));
-    add_evaluator(uid(id, eWorldPropertyThrowIdle), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, CMissile::eThrow));
-    add_evaluator(uid(id, eWorldPropertyThrow), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, CMissile::eThrowEnd));
+    add_evaluator(uid(id, eWorldPropertyHidden), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, gsl::narrow<u32>(CMissile::eHidden)));
+    add_evaluator(uid(id, eWorldPropertyThrowStarted), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, gsl::narrow<u32>(CMissile::eThrowStart)));
+    add_evaluator(uid(id, eWorldPropertyThrowIdle), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, gsl::narrow<u32>(CMissile::eThrow)));
+    add_evaluator(uid(id, eWorldPropertyThrow), xr_new<CObjectPropertyEvaluatorMissile>(missile, m_object, gsl::narrow<u32>(CMissile::eThrowEnd)));
 
     // const properties
     add_evaluator(uid(id, eWorldPropertyDropped), xr_new<CObjectPropertyEvaluatorConst>(false));
@@ -73,7 +73,7 @@ void CObjectHandlerPlanner::add_operators(CMissile* missile)
     add_operator(uid(id, eWorldOperatorIdle), action);
 
     // fire start
-    action = xr_new<CObjectActionFire>(missile, m_object, &m_storage, 0, "throw start");
+    action = xr_new<CObjectActionFire>(missile, m_object, &m_storage, 0u, "throw start");
     add_condition(action, id, eWorldPropertyHidden, false);
     add_condition(action, id, eWorldPropertyThrowStarted, false);
     add_effect(action, id, eWorldPropertyThrowStarted, true);
